Cluster Sequences with VSEARCH

Cluster putative LTR transposons predicted by LTRpred using VSEARCH.

CLUSTpred(file, similarity = 0.9, strand = "both", cores = 1,
  out.name = "CLUSTpred", output = NULL)

Arguments

file
path to predicted LTR transposons generated by LTRpred (in fasta format).

similarity
reject if sequence similarity is lower than this threshold.

strand
cluster using plus or both strands.

cores
number of cores that shall be used for parallel computations.

out.name
name of the output files (*.uc, *.log, *.blast6out).

output
path to a folder in which output shall be stored.

Value

First the following files generated by VSEARCH are stored in the output folder (default: getwd):

*.uc USEARCH cluster format generated by VSEARCH storing the sequence cluster information of the input LTR transposons.

*.log a log file of the VSEARCH run.

*.blast6out BLAST output generated by VSEARCH storing the BLAST hit information of the input LTR transposons.

A USEARCH cluster format (*.uc file extension) table (see read.uc for specifications).

Details

To be able to use this function the VSEARCH command line tool needs to be installed.

References

https://github.com/torognes/vsearch