Query the Dfam database to annotate putative LTRs

Validate or annotate putative LTR transposons that have been predicted using LTRharvest, LTRdigest, or LTRpred.

dfam.query(seq.file, Dfam.db = NULL, eval = 0.001, cores = 1,
  output.folder = getwd())

Arguments

seq.file
file path to the putative LTR transposon sequences in fasta format.

Dfam.db
folder path to the local Dfam database or Dfam.db = "download" in case the Dfam database shall be automatically downloaded before performing query analyses.

eval
E-value threshhold to perform the HMMer search against the Dfam database.

cores
number of cores to use to perform parallel computations.

output.folder
folder path to store the annotation results.

Details

The Dfam database provides a collection of curated transposable element annotations.