Retrieve file names of files genereated by LTRpred

Useful helper function to generate automated file names when crawling LTRpred generated result folders.

get.pred.filenames(LTRpred.folder, ws.wrap = FALSE)

Arguments

LTRpred.folder
path to a LTRpred generated folder.

ws.wrap
shall white space separated path names be wrapped with 'file name'.

Author

Hajk-Georg Drost

Examples

# without wrapped whitespaces
get.pred.filenames("path/to/LTRpred.folder")

#>  [1] "path/to/LTRpred.folder/LTRpred.folder_LTRpred_DataSheet.tsv"                            
#>  [2] "path/to/LTRpred.folder/LTRpred.folder_LTRpred.bed"                                      
#>  [3] "path/to/LTRpred.folder/LTRpred.folder_LTRpred.gff"                                      
#>  [4] "path/to/LTRpred.folder/LTRpred.folder_ORF_prediction_nt.fsa"                            
#>  [5] "path/to/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.blast6out"       
#>  [6] "path/to/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.log"             
#>  [7] "path/to/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.uc"              
#>  [8] "path/to/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_nt.fsa"                    
#>  [9] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_3ltr.fas"      
#> [10] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_5ltr.fas"      
#> [11] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_5ltr.fas"      
#> [12] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder_LTRdigestPrediction.gff" 
#> [13] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_tabout.csv"    
#> [14] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_complete.fas"  
#> [15] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_conditions.csv"
#> [16] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_pbs.fas"       
#> [17] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_ppt.fas"       
#> [18] "path/to/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_pbs.fas"       

# with wrapped whitespaces
get.pred.filenames("path to/LTRpred.folder", ws.wrap = TRUE )

#>  [1] "'path to'/LTRpred.folder/LTRpred.folder_LTRpred_DataSheet.tsv"                            
#>  [2] "'path to'/LTRpred.folder/LTRpred.folder_LTRpred.bed"                                      
#>  [3] "'path to'/LTRpred.folder/LTRpred.folder_LTRpred.gff"                                      
#>  [4] "'path to'/LTRpred.folder/LTRpred.folder_ORF_prediction_nt.fsa"                            
#>  [5] "'path to'/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.blast6out"       
#>  [6] "'path to'/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.log"             
#>  [7] "'path to'/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_CLUSTpred.uc"              
#>  [8] "'path to'/LTRpred.folder/LTRpred.folder-ltrdigest_complete.fas_nt.fsa"                    
#>  [9] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_3ltr.fas"      
#> [10] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_5ltr.fas"      
#> [11] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_5ltr.fas"      
#> [12] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder_LTRdigestPrediction.gff" 
#> [13] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_tabout.csv"    
#> [14] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_complete.fas"  
#> [15] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_conditions.csv"
#> [16] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_pbs.fas"       
#> [17] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_ppt.fas"       
#> [18] "'path to'/LTRpred.folder/LTRpred.folder_ltrdigest/LTRpred.folder-ltrdigest_pbs.fas"