Estimate retrotransposon insertion age in Mya based on 5' and 3' LTR sequence homology

This function implements diverse metrics to roughly estimate the insertion age in Mya based on 5' and 3' LTR sequence homology.

ltr_age_estimation(ltr_seqs_3_prime, ltr_seqs_5_prime, model = "K80",
  mutation_rate = 1.3 * 1e-07)

Arguments

ltr_seqs_3_prime
file path to a fasta file storing the sequences of the respective 3' LTR (e.g. as annotatted by LTRpred).

ltr_seqs_5_prime
file path to a fasta file storing the sequences of the respective 5' LTR (e.g. as annotatted by LTRpred).

model
a model as specified in dist.dna: a character string specifying the evolutionary model to be used; must be one of "raw", "N", "TS", "TV", "JC69", "K80" (the default), "F81", "K81", "F84", "BH87", "T92", "TN93", "GG95", "logdet", "paralin".

mutation_rate
a mutation rate per site per year. For retrotransposons the default is mutation_rate = 1.3 * 10E-8 (Wicker and Keller, 2007).

Examples

# NOT RUN {
 {
# }
# NOT RUN {
# define file path to fasta file storing 3' LTR sequences
# }
# NOT RUN {
ltr_seqs_3_prime <- system.file("Hsapiens_ChrY-ltrdigest_3ltr.fas", package = "LTRpred")
# }# NOT RUN {
ltr_seqs_5_prime <- system.file("Hsapiens_ChrY-ltrdigest_5ltr.fas", package = "LTRpred")
# }# NOT RUN {
# estimate insertion age based on 3' and 5' LTR homology using the K80 model
# }
# NOT RUN {
Hsapiens_ltr_age <- ltr_age_estimation(ltr_seqs_3_prime, ltr_seqs_5_prime)
# }# NOT RUN {
# look at results
# }
# NOT RUN {
Hsapiens_ltr_age
# }