Low level function to detect motifs in strings

Find a specific motif or a sequence of motifs within genomic sequences.

motif.count(seq.file, motif, as.ratio = FALSE)

Arguments

seq.file
path to the genomic sequecne file of interest (e.g. LTR TE seqs predicted by LTRpred).

motif
a character string or vector of strings which shall be counted within each sequence.

as.ratio
shall count values be returned as asbolute frequency (count value) or as relative frequency (percentage).

Author

Hajk-Georg Drost

Examples

# find number of "CG" motifs in predicted LTR transposons
motif.count(seq.file = system.file("LTRseqs.fas",package = "LTRpred"), 
            motif    = "CG")
#>   [1] 245 186 235 116 129 118  98 180 236  75  81 105 127  56 214  77 152 172
#>  [19] 115 581 459 196 129  70 165  84  84 169 124 216 179 263 166 107 133  79
#>  [37] 117 664 113  81 152 102 191  77 146  92 156 116 140 297 104 108 116 204
#>  [55] 173  85 133 260 119 129 291 274 134 113 655 169 224 189  62  67  81 210
#>  [73] 605  75 437 347 178 106  93 146 585  67  96 191 147 159  83 163  79 190
#>  [91] 150 102 109 119 181 135 150 130 216 122 117 131  80 234 188 176  91 187
#> [109] 224 152 124 627 157 367 146 130  66 145 405 265  56 110  63 165  88 300
#> [127] 164  86 110  96  99 239  92  87 164 117 136  83 611 686 576 973 445 112
#> [145] 151  66  91 396  80 111 130 300 191 136 409  62 172 369  84 192 326 190
#> [163] 129 140  88 139 206 203 112  81  82  96 328 229 334 206  89 235 116 165
#> [181] 124 291 100 200  81 177 154 522 154 174 205 124 194 120 397 198 459 615
#> [199] 142  90 173 147 117 108  99  93 125  87 176  77 170 208  64 121 116 318
#> [217] 185 189 112 206 123 124 287 218 103 117 318 564 114 212 304 211 327 223
#> [235] 302 333 401 116 116 129 113 518 204 741 226 137 310 162 120 149  98 225
#> [253] 532  81 169  76 144 218  66  98 452  83 247 418 156  91 254  69 222 128
#> [271]  89 151 134  77 482  94 102  78  95 159 128 126 157 385 385 211  91 145
#> [289] 111 125 247 207 522 117 111 151 186 148 194 125 122 659 660  66 138 101
#> [307] 527 145  88 132  91 147 190 309 115 498 108 187 653 104 257 327 301 147
#> [325] 162 187 134 146 101 240  84 162 168 235  80 114  75 132 111  93 117 132
#> [343] 235 139  78  70  79 188  74  69 350 399  79 131 108  88 181  72 258 122
#> [361] 132 643  74 104 143 137 316 236 368 239 215 177 701  82 176 119 121 176
#> [379] 115  93  73  99 199 171 113 114 219 108 289 119 238 274  76  62 108 207
#> [397] 178 193 116 155  57 592  97 177 310 373  93 308 262 276 162  77 181  71
#> [415]  75  97  77  77 176 130  80 130 128 182 105 128 232 111 116 138  76 143
#> [433] 215 395 214  74 107 199 117 140 100 128 179 109 206 126 117 128 154  78
#> [451] 204 131 191  94 240  96 580 686
            
# find number of "CG" motifs in predicted LTR transposons: rel. frequency
motif.count(seq.file = system.file("LTRseqs.fas",package = "LTRpred"), 
            motif    = "CG",
            as.ratio = TRUE)             
#>     1_CHROMOSOME_dumped__149500_155909     1_CHROMOSOME_dumped__249531_259127 
#>                             0.03822153                             0.01938106 
#>     1_CHROMOSOME_dumped__428932_438092   1_CHROMOSOME_dumped__2659374_2663971 
#>                             0.02565222                             0.02522836 
#>   1_CHROMOSOME_dumped__3357954_3363283   1_CHROMOSOME_dumped__3780765_3785720 
#>                             0.02420263                             0.02380952 
#>   1_CHROMOSOME_dumped__3907118_3911673   1_CHROMOSOME_dumped__4081628_4091086 
#>                             0.02151010                             0.01902950 
#>   1_CHROMOSOME_dumped__4119030_4126617   1_CHROMOSOME_dumped__4405997_4411118 
#>                             0.03110174                             0.01464272 
#>   1_CHROMOSOME_dumped__5691759_5696703   1_CHROMOSOME_dumped__5897842_5902839 
#>                             0.01638018                             0.02100840 
#>   1_CHROMOSOME_dumped__5920835_5927113   1_CHROMOSOME_dumped__6009304_6014076 
#>                             0.02022615                             0.01173266 
#>   1_CHROMOSOME_dumped__6191403_6197517   1_CHROMOSOME_dumped__6538418_6543536 
#>                             0.03499591                             0.01504200 
#>   1_CHROMOSOME_dumped__6719067_6724311   1_CHROMOSOME_dumped__6767255_6773967 
#>                             0.02897998                             0.02562193 
#>   1_CHROMOSOME_dumped__6825176_6830220   1_CHROMOSOME_dumped__7025897_7047012 
#>                             0.02279485                             0.02751468 
#>   1_CHROMOSOME_dumped__8445394_8466734   1_CHROMOSOME_dumped__8578711_8588761 
#>                             0.02150790                             0.01950055 
#>   1_CHROMOSOME_dumped__8643650_8649288   1_CHROMOSOME_dumped__9368344_9373311 
#>                             0.02287640                             0.01409018 
#>   1_CHROMOSOME_dumped__9476002_9480634 1_CHROMOSOME_dumped__10121579_10125768 
#>                             0.03561407                             0.02004773 
#> 1_CHROMOSOME_dumped__10696719_10701589 1_CHROMOSOME_dumped__11125145_11133104 
#>                             0.01724492                             0.02123116 
#> 1_CHROMOSOME_dumped__11148084_11152830 1_CHROMOSOME_dumped__11284903_11295619 
#>                             0.02612176                             0.02015489 
#> 1_CHROMOSOME_dumped__11339006_11344662 1_CHROMOSOME_dumped__11346530_11355061 
#>                             0.03164221                             0.03082513 
#> 1_CHROMOSOME_dumped__11697950_11705480 1_CHROMOSOME_dumped__11789447_11794074 
#>                             0.02204223                             0.02312014 
#> 1_CHROMOSOME_dumped__11890866_11900756 1_CHROMOSOME_dumped__12194788_12199420 
#>                             0.01344657                             0.01705159 
#> 1_CHROMOSOME_dumped__12541821_12546900 1_CHROMOSOME_dumped__12664511_12673933 
#>                             0.02303150                             0.07046588 
#> 1_CHROMOSOME_dumped__12871512_12878779 1_CHROMOSOME_dumped__12969459_12975236 
#>                             0.01554761                             0.01401869 
#> 1_CHROMOSOME_dumped__13324761_13329591 1_CHROMOSOME_dumped__13334465_13339870 
#>                             0.03146347                             0.01886792 
#> 1_CHROMOSOME_dumped__13388869_13396650 1_CHROMOSOME_dumped__13398499_13402662 
#>                             0.02454382                             0.01849183 
#> 1_CHROMOSOME_dumped__13439443_13448531 1_CHROMOSOME_dumped__13559843_13564825 
#>                             0.01606337                             0.01846277 
#> 1_CHROMOSOME_dumped__13593058_13599710 1_CHROMOSOME_dumped__13676498_13682108 
#>                             0.02344807                             0.02067368 
#> 1_CHROMOSOME_dumped__13818567_13823889 1_CHROMOSOME_dumped__13883095_13890935 
#>                             0.02630096                             0.03787782 
#> 1_CHROMOSOME_dumped__13912841_13919010 1_CHROMOSOME_dumped__14019278_14024519 
#>                             0.01685575                             0.02060282 
#> 1_CHROMOSOME_dumped__14179555_14183988 1_CHROMOSOME_dumped__14214740_14225657 
#>                             0.02616148                             0.01868474 
#> 1_CHROMOSOME_dumped__14282118_14286791 1_CHROMOSOME_dumped__14745962_14750821 
#>                             0.03701326                             0.01748971 
#> 1_CHROMOSOME_dumped__15610250_15615956 1_CHROMOSOME_dumped__16012601_16019700 
#>                             0.02330471                             0.03661972 
#> 1_CHROMOSOME_dumped__16117817_16124067 1_CHROMOSOME_dumped__16176836_16181678 
#>                             0.01903695                             0.02663638 
#> 1_CHROMOSOME_dumped__16387793_16398190 1_CHROMOSOME_dumped__16537640_16548760 
#>                             0.02798615                             0.02463807 
#> 1_CHROMOSOME_dumped__16551721_16557069 1_CHROMOSOME_dumped__16569870_16575240 
#>                             0.02505141                             0.02103891 
#> 1_CHROMOSOME_dumped__16626036_16637234 1_CHROMOSOME_dumped__16638793_16645558 
#>                             0.05848736                             0.02497783 
#> 1_CHROMOSOME_dumped__16674908_16683328 1_CHROMOSOME_dumped__16853317_16858040 
#>                             0.02660017                             0.04000847 
#> 1_CHROMOSOME_dumped__17023916_17029257 1_CHROMOSOME_dumped__17054990_17060023 
#>                             0.01160614                             0.01330950 
#> 1_CHROMOSOME_dumped__17156581_17161059 1_CHROMOSOME_dumped__17247721_17252580 
#>                             0.01808439                             0.04320988 
#> 1_CHROMOSOME_dumped__17255285_17265290 1_CHROMOSOME_dumped__17520596_17525126 
#>                             0.06046372                             0.01655264 
#> 1_CHROMOSOME_dumped__17528667_17536139 1_CHROMOSOME_dumped__17574998_17592212 
#>                             0.05847718                             0.02015684 
#> 1_CHROMOSOME_dumped__17606460_17612336 1_CHROMOSOME_dumped__18198315_18206598 
#>                             0.03028756                             0.01279575 
#> 1_CHROMOSOME_dumped__18552993_18557995 1_CHROMOSOME_dumped__18639849_18648168 
#>                             0.01858885                             0.01754808 
#> 1_CHROMOSOME_dumped__18962095_18971496 1_CHROMOSOME_dumped__18996885_19001831 
#>                             0.06222080                             0.01354356 
#> 1_CHROMOSOME_dumped__20085162_20090239 1_CHROMOSOME_dumped__21244323_21249201 
#>                             0.01890508                             0.03914737 
#> 1_CHROMOSOME_dumped__21318903_21325509 1_CHROMOSOME_dumped__21346162_21351041 
#>                             0.02224913                             0.03258197 
#> 1_CHROMOSOME_dumped__21438251_21442649 1_CHROMOSOME_dumped__21930876_21937598 
#>                             0.01886792                             0.02424513 
#> 1_CHROMOSOME_dumped__21947272_21951934 1_CHROMOSOME_dumped__22100643_22105329 
#>                             0.01694188                             0.04053766 
#> 1_CHROMOSOME_dumped__22272951_22280403 1_CHROMOSOME_dumped__22619338_22624555 
#>                             0.02012612                             0.01954772 
#> 1_CHROMOSOME_dumped__22831590_22837366 1_CHROMOSOME_dumped__22941592_22946917 
#>                             0.01886792                             0.02234322 
#> 1_CHROMOSOME_dumped__23000655_23008290 1_CHROMOSOME_dumped__23118242_23123215 
#>                             0.02370351                             0.02714113 
#> 1_CHROMOSOME_dumped__23569996_23575851 1_CHROMOSOME_dumped__23646949_23653338 
#>                             0.02561475                             0.02034429 
#> 1_CHROMOSOME_dumped__24506842_24514619 1_CHROMOSOME_dumped__24873949_24879663 
#>                             0.02777064                             0.02134733 
#> 1_CHROMOSOME_dumped__24938566_24945901 1_CHROMOSOME_dumped__25265020_25271109 
#>                             0.01594875                             0.02151067 
#> 1_CHROMOSOME_dumped__26366608_26371676 1_CHROMOSOME_dumped__27199829_27210847 
#>                             0.01578221                             0.02123605 
#> 1_CHROMOSOME_dumped__27829437_27835759 1_CHROMOSOME_dumped__27930411_27936476 
#>                             0.02973272                             0.02901418 
#> 1_CHROMOSOME_dumped__28501879_28506082 1_CHROMOSOME_dumped__29587846_29595895 
#>                             0.02164605                             0.02322981 
#> 1_CHROMOSOME_dumped__29642083_29647526 1_CHROMOSOME_dumped__29700884_29707698 
#>                             0.04114622                             0.02230374 
#> 1_CHROMOSOME_dumped__30419866_30424328        2_CHROMOSOME_dumped__2080_10521 
#>                             0.02778400                             0.07427150 
#>     2_CHROMOSOME_dumped__167975_174024     2_CHROMOSOME_dumped__593307_610013 
#>                             0.02595041                             0.02196684 
#>     2_CHROMOSOME_dumped__611690_617410     2_CHROMOSOME_dumped__775782_782412 
#>                             0.02552001                             0.01960489 
#>     2_CHROMOSOME_dumped__909733_914339     2_CHROMOSOME_dumped__959966_964731 
#>                             0.01432603                             0.03042384 
#>     2_CHROMOSOME_dumped__975729_990317   2_CHROMOSOME_dumped__1235983_1250415 
#>                             0.02776064                             0.01836070 
#>   2_CHROMOSOME_dumped__1292790_1297013   2_CHROMOSOME_dumped__1388832_1393901 
#>                             0.01325758                             0.02169625 
#>   2_CHROMOSOME_dumped__1395261_1399523   2_CHROMOSOME_dumped__1523655_1531024 
#>                             0.01477833                             0.02238806 
#>   2_CHROMOSOME_dumped__1557140_1562062   2_CHROMOSOME_dumped__1629671_1637490 
#>                             0.01787528                             0.03836317 
#>   2_CHROMOSOME_dumped__1882607_1887821   2_CHROMOSOME_dumped__1970254_1974960 
#>                             0.03144775                             0.01827066 
#>   2_CHROMOSOME_dumped__2075307_2080607   2_CHROMOSOME_dumped__2274202_2280048 
#>                             0.02075080                             0.01641868 
#>   2_CHROMOSOME_dumped__2302464_2307170   2_CHROMOSOME_dumped__2745600_2752272 
#>                             0.02103250                             0.03581597 
#>   2_CHROMOSOME_dumped__2844749_2849616   2_CHROMOSOME_dumped__2858551_2864192 
#>                             0.01889893                             0.01542006 
#>   2_CHROMOSOME_dumped__2902330_2907117   2_CHROMOSOME_dumped__2911997_2919582 
#>                             0.03425230                             0.01542315 
#>   2_CHROMOSOME_dumped__3083903_3089468   2_CHROMOSOME_dumped__3160549_3166281 
#>                             0.02443406                             0.01447759 
#>   2_CHROMOSOME_dumped__3221681_3238915   2_CHROMOSOME_dumped__3270279_3289109 
#>                             0.03545112                             0.03642929 
#>   2_CHROMOSOME_dumped__3367929_3386289   2_CHROMOSOME_dumped__3426174_3451000 
#>                             0.03137084                             0.03919120 
#>   2_CHROMOSOME_dumped__3539977_3559406   2_CHROMOSOME_dumped__4080728_4088155 
#>                             0.02290273                             0.01507808 
#>   2_CHROMOSOME_dumped__4090709_4096103   2_CHROMOSOME_dumped__4110947_4115696 
#>                             0.02798888                             0.01389474 
#>   2_CHROMOSOME_dumped__4116023_4120740   2_CHROMOSOME_dumped__4161050_4173975 
#>                             0.01928783                             0.03063593 
#>   2_CHROMOSOME_dumped__4230743_4236513   2_CHROMOSOME_dumped__4408338_4413617 
#>                             0.01386242                             0.02102273 
#>   2_CHROMOSOME_dumped__4542920_4547698   2_CHROMOSOME_dumped__4558517_4566125 
#>                             0.02720234                             0.03942699 
#>   2_CHROMOSOME_dumped__4682601_4690775   2_CHROMOSOME_dumped__4693482_4698777 
#>                             0.02336391                             0.02567976 
#>   2_CHROMOSOME_dumped__4737963_4753533   2_CHROMOSOME_dumped__4775685_4780521 
#>                             0.02626678                             0.01281786 
#>   2_CHROMOSOME_dumped__4818686_4826103   2_CHROMOSOME_dumped__5040989_5048393 
#>                             0.02318684                             0.04983120 
#>   2_CHROMOSOME_dumped__5073848_5078420   2_CHROMOSOME_dumped__5394710_5399650 
#>                             0.01836869                             0.03885853 
#>   2_CHROMOSOME_dumped__6023868_6032902   2_CHROMOSOME_dumped__6256496_6260935 
#>                             0.03608190                             0.04279279 
#>   2_CHROMOSOME_dumped__6422229_6427295   2_CHROMOSOME_dumped__6819006_6823730 
#>                             0.02545885                             0.02962963 
#>   2_CHROMOSOME_dumped__6837314_6843322   2_CHROMOSOME_dumped__6894106_6899938 
#>                             0.01464470                             0.02382993 
#>   2_CHROMOSOME_dumped__7157625_7166649   2_CHROMOSOME_dumped__7224084_7229138 
#>                             0.02282548                             0.04015826 
#>   2_CHROMOSOME_dumped__7373364_7379268   2_CHROMOSOME_dumped__7448762_7454290 
#>                             0.01896698                             0.01465003 
#>   2_CHROMOSOME_dumped__8697239_8702070   2_CHROMOSOME_dumped__8816049_8820894 
#>                             0.01697020                             0.01981015 
#>   2_CHROMOSOME_dumped__8873017_8891285   2_CHROMOSOME_dumped__9187492_9195280 
#>                             0.01795391                             0.02940044 
#>   2_CHROMOSOME_dumped__9306572_9319034   2_CHROMOSOME_dumped__9927681_9932559 
#>                             0.02679933                             0.04222177 
#> 2_CHROMOSOME_dumped__10464648_10471481 2_CHROMOSOME_dumped__10483666_10492018 
#>                             0.01302312                             0.02813360 
#> 2_CHROMOSOME_dumped__10637411_10647335 2_CHROMOSOME_dumped__11435732_11443090 
#>                             0.01168766                             0.02242152 
#> 2_CHROMOSOME_dumped__11721105_11726615 2_CHROMOSOME_dumped__12445635_12457164 
#>                             0.02250045                             0.02523851 
#> 2_CHROMOSOME_dumped__12620243_12625077 2_CHROMOSOME_dumped__13508554_13515938 
#>                             0.02068252                             0.02708192 
#> 2_CHROMOSOME_dumped__13664992_13669870 2_CHROMOSOME_dumped__14026300_14032819 
#>                             0.01660176                             0.02714724 
#> 2_CHROMOSOME_dumped__14406571_14412855 2_CHROMOSOME_dumped__16280712_16302420 
#>                             0.02450278                             0.02404533 
#> 2_CHROMOSOME_dumped__17136148_17141211 2_CHROMOSOME_dumped__17252349_17258537 
#>                             0.03041074                             0.02811440 
#> 2_CHROMOSOME_dumped__17270937_17278587 2_CHROMOSOME_dumped__17676479_17684201 
#>                             0.02679388                             0.01605594 
#> 2_CHROMOSOME_dumped__18347446_18355355 2_CHROMOSOME_dumped__18659187_18664902 
#>                             0.02452592                             0.02099370 
#>     3_CHROMOSOME_dumped__780904_798849   3_CHROMOSOME_dumped__1526151_1535730 
#>                             0.02212192                             0.02066806 
#>   3_CHROMOSOME_dumped__1657651_1681989   3_CHROMOSOME_dumped__3110405_3124906 
#>                             0.01885862                             0.04240794 
#>   3_CHROMOSOME_dumped__3807306_3812643   3_CHROMOSOME_dumped__6188741_6193283 
#>                             0.02660172                             0.01981070 
#>   3_CHROMOSOME_dumped__6372409_6378607   3_CHROMOSOME_dumped__7354085_7359059 
#>                             0.02790773                             0.02954774 
#>   3_CHROMOSOME_dumped__7728276_7733244   3_CHROMOSOME_dumped__7760615_7765156 
#>                             0.02354599                             0.02377807 
#>   3_CHROMOSOME_dumped__8046323_8051045   3_CHROMOSOME_dumped__8537355_8543559 
#>                             0.02096125                             0.01498791 
#>   3_CHROMOSOME_dumped__8632393_8638600   3_CHROMOSOME_dumped__9226633_9231377 
#>                             0.02013531                             0.01833509 
#>   3_CHROMOSOME_dumped__9243115_9251762   3_CHROMOSOME_dumped__9387852_9393276 
#>                             0.02035153                             0.01419355 
#>   3_CHROMOSOME_dumped__9404664_9410413   3_CHROMOSOME_dumped__9604460_9613216 
#>                             0.02956522                             0.02375243 
#>   3_CHROMOSOME_dumped__9692211_9697532 3_CHROMOSOME_dumped__10494353_10500970 
#>                             0.01202555                             0.01828347 
#> 3_CHROMOSOME_dumped__10577158_10581825 3_CHROMOSOME_dumped__10711160_10725823 
#>                             0.02485004                             0.02168576 
#> 3_CHROMOSOME_dumped__11016111_11020804 3_CHROMOSOME_dumped__11122707_11127675 
#>                             0.03941202                             0.03803582 
#> 3_CHROMOSOME_dumped__11387801_11394402 3_CHROMOSOME_dumped__11652697_11657743 
#>                             0.01696456                             0.04081633 
#> 3_CHROMOSOME_dumped__11664969_11669787 3_CHROMOSOME_dumped__11853471_11859849 
#>                             0.02552397                             0.01943878 
#> 3_CHROMOSOME_dumped__12056081_12071267 3_CHROMOSOME_dumped__12275508_12287360 
#>                             0.01889774                             0.01839197 
#> 3_CHROMOSOME_dumped__12303477_12308746 3_CHROMOSOME_dumped__12314647_12319070 
#>                             0.01954459                             0.02644665 
#> 3_CHROMOSOME_dumped__12478086_12486327 3_CHROMOSOME_dumped__12491757_12502106 
#>                             0.03858287                             0.05449275 
#> 3_CHROMOSOME_dumped__12502664_12508397 3_CHROMOSOME_dumped__12549509_12555919 
#>                             0.01988141                             0.03306816 
#> 3_CHROMOSOME_dumped__12613331_12620941 3_CHROMOSOME_dumped__12636500_12644002 
#>                             0.03994219                             0.02812208 
#> 3_CHROMOSOME_dumped__12674274_12684173 3_CHROMOSOME_dumped__12695928_12704543 
#>                             0.03303030                             0.02588208 
#> 3_CHROMOSOME_dumped__12960389_12968468 3_CHROMOSOME_dumped__13087350_13092790 
#>                             0.03737624                             0.06120198 
#> 3_CHROMOSOME_dumped__13096163_13112097 3_CHROMOSOME_dumped__13318010_13323173 
#>                             0.02516473                             0.02246321 
#> 3_CHROMOSOME_dumped__13334612_13342566 3_CHROMOSOME_dumped__13360109_13368309 
#>                             0.01458202                             0.01572979 
#> 3_CHROMOSOME_dumped__13369174_13374107 3_CHROMOSOME_dumped__13440036_13449242 
#>                             0.02290231                             0.05626154 
#> 3_CHROMOSOME_dumped__14073146_14081964 3_CHROMOSOME_dumped__14117876_14141561 
#>                             0.02313187                             0.03128430 
#> 3_CHROMOSOME_dumped__14272824_14285579 3_CHROMOSOME_dumped__14329874_14334136 
#>                             0.01771715                             0.03213699 
#> 3_CHROMOSOME_dumped__14504089_14511689 3_CHROMOSOME_dumped__14579222_14584089 
#>                             0.04078411                             0.03327855 
#> 3_CHROMOSOME_dumped__14608928_14615360 3_CHROMOSOME_dumped__14616107_14623209 
#>                             0.01865382                             0.02097705 
#> 3_CHROMOSOME_dumped__14671141_14675856 3_CHROMOSOME_dumped__14729381_14734607 
#>                             0.02078032                             0.04304572 
#> 3_CHROMOSOME_dumped__14762490_14785730 3_CHROMOSOME_dumped__14860990_14866191 
#>                             0.02289058                             0.01557093 
#> 3_CHROMOSOME_dumped__14890955_14900038 3_CHROMOSOME_dumped__14956986_14962254 
#>                             0.01860414                             0.01442399 
#> 3_CHROMOSOME_dumped__15540529_15549180 3_CHROMOSOME_dumped__15591669_15600822 
#>                             0.01664355                             0.02381473 
#> 3_CHROMOSOME_dumped__15621872_15626854 3_CHROMOSOME_dumped__15768786_15773747 
#>                             0.01324503                             0.01975010 
#> 3_CHROMOSOME_dumped__15779153_15801235 3_CHROMOSOME_dumped__15919026_15924075 
#>                             0.02046823                             0.01643564 
#> 3_CHROMOSOME_dumped__16172247_16183071 3_CHROMOSOME_dumped__16243296_16256896 
#>                             0.02281755                             0.03073303 
#> 3_CHROMOSOME_dumped__16809187_16814152 3_CHROMOSOME_dumped__16826414_16832203 
#>                             0.03141361                             0.01571675 
#> 3_CHROMOSOME_dumped__16885889_16894261 3_CHROMOSOME_dumped__16957465_16962731 
#>                             0.03033560                             0.01310044 
#> 3_CHROMOSOME_dumped__17031074_17041106 3_CHROMOSOME_dumped__17235380_17239718 
#>                             0.02212698                             0.02949988 
#> 3_CHROMOSOME_dumped__17354612_17359635 3_CHROMOSOME_dumped__17403584_17414141 
#>                             0.01771497                             0.01430195 
#> 3_CHROMOSOME_dumped__17447389_17451847 3_CHROMOSOME_dumped__17978754_17983948 
#>                             0.03005158                             0.01482194 
#> 3_CHROMOSOME_dumped__18151296_18168081 3_CHROMOSOME_dumped__18169847_18174127 
#>                             0.02871440                             0.02195749 
#> 3_CHROMOSOME_dumped__18536569_18541113 3_CHROMOSOME_dumped__18734248_18739393 
#>                             0.02244224                             0.01515740 
#> 3_CHROMOSOME_dumped__18791152_18796089 3_CHROMOSOME_dumped__19290470_19296623 
#>                             0.01923856                             0.02583685 
#> 3_CHROMOSOME_dumped__19759349_19765040 3_CHROMOSOME_dumped__20304946_20310167 
#>                             0.02248770                             0.02412869 
#> 3_CHROMOSOME_dumped__21485038_21491983 3_CHROMOSOME_dumped__21576037_21593706 
#>                             0.02260294                             0.02178834 
#> 3_CHROMOSOME_dumped__21642929_21660064 3_CHROMOSOME_dumped__21894821_21904623 
#>                             0.02246732                             0.02152402 
#> 3_CHROMOSOME_dumped__21971055_21975268 3_CHROMOSOME_dumped__22015001_22020670 
#>                             0.02159468                             0.02557319 
#> 3_CHROMOSOME_dumped__22059535_22064329 3_CHROMOSOME_dumped__22232190_22236820 
#>                             0.02314911                             0.02699201 
#> 3_CHROMOSOME_dumped__22237229_22242466 3_CHROMOSOME_dumped__22334467_22341333 
#>                             0.04715540                             0.03014417 
#> 3_CHROMOSOME_dumped__22513587_22521224 3_CHROMOSOME_dumped__22695566_22700521 
#>                             0.06834250                             0.02360775 
#> 3_CHROMOSOME_dumped__23100313_23104853     4_CHROMOSOME_dumped__178046_186886 
#>                             0.02444396                             0.01707952 
#>   4_CHROMOSOME_dumped__1340159_1351874   4_CHROMOSOME_dumped__1473821_1478344 
#>                             0.01587573                             0.03271441 
#>   4_CHROMOSOME_dumped__1619054_1624381   4_CHROMOSOME_dumped__1749269_1755807 
#>                             0.03641141                             0.01911607 
#>   4_CHROMOSOME_dumped__1814108_1818633   4_CHROMOSOME_dumped__1943797_1955321 
#>                             0.02695537                             0.05718004 
#>   4_CHROMOSOME_dumped__2001193_2010885   4_CHROMOSOME_dumped__2064089_2068818 
#>                             0.06809037                             0.01395349 
#>   4_CHROMOSOME_dumped__2188978_2194314   4_CHROMOSOME_dumped__2198167_2203996 
#>                             0.02585722                             0.01732419 
#>   4_CHROMOSOME_dumped__2243443_2261156   4_CHROMOSOME_dumped__2279948_2286705 
#>                             0.02975048                             0.02145605 
#>   4_CHROMOSOME_dumped__2302657_2307843   4_CHROMOSOME_dumped__2442384_2446881 
#>                             0.01696549                             0.02934638 
#>   4_CHROMOSOME_dumped__2597020_2602180   4_CHROMOSOME_dumped__2615229_2622565 
#>                             0.01763224                             0.02003544 
#>   4_CHROMOSOME_dumped__2763293_2771367   4_CHROMOSOME_dumped__2867762_2875913 
#>                             0.02352941                             0.03790481 
#>   4_CHROMOSOME_dumped__2989830_2994788   4_CHROMOSOME_dumped__3704587_3716723 
#>                             0.02319016                             0.04103156 
#>   4_CHROMOSOME_dumped__3881650_3886431   4_CHROMOSOME_dumped__3907466_3915200 
#>                             0.02258469                             0.02417582 
#>   4_CHROMOSOME_dumped__4464332_4482516   4_CHROMOSOME_dumped__4623486_4628226 
#>                             0.03590872                             0.02193630 
#>   4_CHROMOSOME_dumped__4677141_4684195   4_CHROMOSOME_dumped__4958195_4969598 
#>                             0.03642807                             0.02867415 
#>   4_CHROMOSOME_dumped__5127101_5135153   4_CHROMOSOME_dumped__5291921_5296999 
#>                             0.03737737                             0.02894271 
#>   4_CHROMOSOME_dumped__5591440_5596425   4_CHROMOSOME_dumped__5822426_5827300 
#>                             0.03249097                             0.03835897 
#>   4_CHROMOSOME_dumped__5987008_5991992   4_CHROMOSOME_dumped__6588726_6593808 
#>                             0.02688064                             0.02872319 
#>   4_CHROMOSOME_dumped__6594136_6598635   4_CHROMOSOME_dumped__6659028_6672586 
#>                             0.02244444                             0.01770042 
#>   4_CHROMOSOME_dumped__6723347_6728791   4_CHROMOSOME_dumped__6826682_6833259 
#>                             0.01542700                             0.02462755 
#>   4_CHROMOSOME_dumped__6914662_6919598   4_CHROMOSOME_dumped__7298372_7305107 
#>                             0.03402876                             0.03488717 
#>   4_CHROMOSOME_dumped__7359462_7364591   4_CHROMOSOME_dumped__7522904_7527330 
#>                             0.01559454                             0.02575107 
#>   4_CHROMOSOME_dumped__7816826_7821792   4_CHROMOSOME_dumped__8086223_8092247 
#>                             0.01509966                             0.02190871 
#>   4_CHROMOSOME_dumped__8202631_8206965   4_CHROMOSOME_dumped__8313595_8319066 
#>                             0.02560554                             0.01699561 
#>   4_CHROMOSOME_dumped__8589987_8597164   4_CHROMOSOME_dumped__8852913_8857290 
#>                             0.01629980                             0.03015075 
#>   4_CHROMOSOME_dumped__9163509_9172988   4_CHROMOSOME_dumped__9182952_9188015 
#>                             0.02478903                             0.02744866 
#>   4_CHROMOSOME_dumped__9570498_9575604   4_CHROMOSOME_dumped__9732605_9737731 
#>                             0.01527315                             0.01365321 
#>   4_CHROMOSOME_dumped__9908989_9914046   4_CHROMOSOME_dumped__9950615_9958242 
#>                             0.01561882                             0.02464604 
#> 4_CHROMOSOME_dumped__10899527_10904693 4_CHROMOSOME_dumped__10992996_10998058 
#>                             0.01432166                             0.01362828 
#> 4_CHROMOSOME_dumped__11040514_11046715 4_CHROMOSOME_dumped__11206541_11211219 
#>                             0.05643341                             0.08527463 
#> 4_CHROMOSOME_dumped__11481824_11486753 4_CHROMOSOME_dumped__11703299_11707622 
#>                             0.01602434                             0.03029602 
#> 4_CHROMOSOME_dumped__11734522_11739041 4_CHROMOSOME_dumped__12211784_12218169 
#>                             0.02389381                             0.01378014 
#> 4_CHROMOSOME_dumped__13336302_13341644 4_CHROMOSOME_dumped__14258825_14263830 
#>                             0.03387610                             0.01438274 
#> 4_CHROMOSOME_dumped__14503006_14512023 4_CHROMOSOME_dumped__14684005_14688751 
#>                             0.02860945                             0.02570044 
#> 4_CHROMOSOME_dumped__15027814_15033127 4_CHROMOSOME_dumped__15551901_15565729 
#>                             0.02484005                             0.04649649 
#> 4_CHROMOSOME_dumped__15840427_15844699 4_CHROMOSOME_dumped__16320863_16325954 
#>                             0.01731804                             0.02042419 
#> 4_CHROMOSOME_dumped__17225299_17230933 4_CHROMOSOME_dumped__17712151_17716968 
#>                             0.02537711                             0.02843504 
#> 4_CHROMOSOME_dumped__18381613_18390512       5_CHROMOSOME_dumped__64772_71199 
#>                             0.03550562                             0.03671437 
#>     5_CHROMOSOME_dumped__211336_219089   5_CHROMOSOME_dumped__1331310_1337190 
#>                             0.04745938                             0.04063935 
#>   5_CHROMOSOME_dumped__1961367_1970365   5_CHROMOSOME_dumped__2373775_2379810 
#>                             0.02389154                             0.02932406 
#>   5_CHROMOSOME_dumped__2380969_2405489   5_CHROMOSOME_dumped__2510278_2515356 
#>                             0.02858774                             0.01614491 
#>   5_CHROMOSOME_dumped__3247646_3253022   5_CHROMOSOME_dumped__3904675_3909862 
#>                             0.03273201                             0.02293755 
#>   5_CHROMOSOME_dumped__4208083_4213084   5_CHROMOSOME_dumped__5603823_5608721 
#>                             0.02419032                             0.03592570 
#>   5_CHROMOSOME_dumped__6064597_6068747   5_CHROMOSOME_dumped__6403057_6408054 
#>                             0.02770417                             0.01860744 
#>   5_CHROMOSOME_dumped__6433629_6438777   5_CHROMOSOME_dumped__6468251_6472831 
#>                             0.01417751                             0.02161100 
#>   5_CHROMOSOME_dumped__7170326_7177335   5_CHROMOSOME_dumped__7210794_7219564 
#>                             0.02838802                             0.01949607 
#>   5_CHROMOSOME_dumped__7505453_7511007   5_CHROMOSOME_dumped__7511425_7516841 
#>                             0.02034203                             0.02104486 
#>   5_CHROMOSOME_dumped__7820047_7826629   5_CHROMOSOME_dumped__7858641_7864260 
#>                             0.03326751                             0.01921708 
#>   5_CHROMOSOME_dumped__8204176_8218846   5_CHROMOSOME_dumped__8245442_8252558 
#>                             0.01969873                             0.01672053 
#>   5_CHROMOSOME_dumped__8927997_8934328   5_CHROMOSOME_dumped__9181894_9189365 
#>                             0.03758686                             0.03667024 
#>   5_CHROMOSOME_dumped__9381694_9386622   5_CHROMOSOME_dumped__9949175_9954432 
#>                             0.01541895                             0.01179156 
#> 5_CHROMOSOME_dumped__10026369_10030633 5_CHROMOSOME_dumped__10128753_10133863 
#>                             0.02532239                             0.04050088 
#> 5_CHROMOSOME_dumped__10359352_10364274 5_CHROMOSOME_dumped__10638480_10643372 
#>                             0.03615681                             0.03944410 
#> 5_CHROMOSOME_dumped__10808404_10812940 5_CHROMOSOME_dumped__11159506_11166120 
#>                             0.02556756                             0.02343159 
#> 5_CHROMOSOME_dumped__11491567_11496131 5_CHROMOSOME_dumped__11503488_11512862 
#>                             0.01248631                             0.06314667 
#> 5_CHROMOSOME_dumped__11515554_11521087 5_CHROMOSOME_dumped__11621561_11635005 
#>                             0.01752801                             0.01316475 
#> 5_CHROMOSOME_dumped__11980554_11994151 5_CHROMOSOME_dumped__12189060_12194348 
#>                             0.02279747                             0.07052373 
#> 5_CHROMOSOME_dumped__13253687_13258453 5_CHROMOSOME_dumped__13261021_13272510 
#>                             0.01950913                             0.02680592 
#> 5_CHROMOSOME_dumped__13344204_13354020 5_CHROMOSOME_dumped__13607605_13621821 
#>                             0.02668840                             0.01941338 
#> 5_CHROMOSOME_dumped__13712934_13721136 5_CHROMOSOME_dumped__13736102_13741134 
#>                             0.01974887                             0.01529903 
#> 5_CHROMOSOME_dumped__13948926_13953982 5_CHROMOSOME_dumped__14083253_14088531 
#>                             0.03579197                             0.01344952 
#> 5_CHROMOSOME_dumped__14113154_14118424 5_CHROMOSOME_dumped__14203867_14209072 
#>                             0.01422880                             0.01863235 
#> 5_CHROMOSOME_dumped__14409576_14414448 5_CHROMOSOME_dumped__14468876_14473748 
#>                             0.01580135                             0.01580135 
#> 5_CHROMOSOME_dumped__14756546_14764119 5_CHROMOSOME_dumped__14873452_14880646 
#>                             0.02323739                             0.01806810 
#> 5_CHROMOSOME_dumped__15175722_15180975 5_CHROMOSOME_dumped__15277978_15284587 
#>                             0.01522649                             0.01966717 
#> 5_CHROMOSOME_dumped__15320410_15326088 5_CHROMOSOME_dumped__15377478_15384917 
#>                             0.02253918                             0.02446237 
#> 5_CHROMOSOME_dumped__15411867_15416287 5_CHROMOSOME_dumped__15646198_15651172 
#>                             0.02375028                             0.02572864 
#> 5_CHROMOSOME_dumped__15813256_15821248 5_CHROMOSOME_dumped__16509654_16515435 
#>                             0.02902540                             0.01919751 
#> 5_CHROMOSOME_dumped__16615185_16620067 5_CHROMOSOME_dumped__16882616_16889700 
#>                             0.02375589                             0.01947777 
#> 5_CHROMOSOME_dumped__17090770_17095675 5_CHROMOSOME_dumped__17394157_17400642 
#>                             0.01549124                             0.02204749 
#> 5_CHROMOSOME_dumped__17596764_17605847 5_CHROMOSOME_dumped__18119049_18132012 
#>                             0.02366799                             0.03046899 
#> 5_CHROMOSOME_dumped__18142156_18146897 5_CHROMOSOME_dumped__18233616_18238203 
#>                             0.04512864                             0.01612903 
#> 5_CHROMOSOME_dumped__18490606_18495233 5_CHROMOSOME_dumped__18511248_18518552 
#>                             0.02312014                             0.02724162 
#> 5_CHROMOSOME_dumped__18551370_18557055 5_CHROMOSOME_dumped__18673108_18678357 
#>                             0.02057686                             0.02666667 
#> 5_CHROMOSOME_dumped__18933102_18938947 5_CHROMOSOME_dumped__19345986_19351107 
#>                             0.01710571                             0.02499024 
#> 5_CHROMOSOME_dumped__19700784_19710576 5_CHROMOSOME_dumped__19987111_19992585 
#>                             0.01827836                             0.01990868 
#> 5_CHROMOSOME_dumped__20391919_20401644 5_CHROMOSOME_dumped__21384430_21390116 
#>                             0.02118034                             0.02215579 
#> 5_CHROMOSOME_dumped__21800243_21805138 5_CHROMOSOME_dumped__21915762_21920415 
#>                             0.02389706                             0.02750322 
#> 5_CHROMOSOME_dumped__22002013_22006896 5_CHROMOSOME_dumped__22911048_22916161 
#>                             0.03153153                             0.01525225 
#> 5_CHROMOSOME_dumped__24129974_24138730 5_CHROMOSOME_dumped__24526077_24530973 
#>                             0.02329565                             0.02675107 
#> 5_CHROMOSOME_dumped__25509002_25517288 5_CHROMOSOME_dumped__25992308_25997851 
#>                             0.02304815                             0.01695527 
#> 5_CHROMOSOME_dumped__26207873_26219365 5_CHROMOSOME_dumped__26587592_26591987 
#>                             0.02088228                             0.02183803 
#>     mitochondria_CHROMOS_111563_129916     mitochondria_CHROMOS_241879_260732 
#>                             0.03160074                             0.03638485