Plot a predicted retrotransposons along the chromosomes

LTR retrotransposons predicted with LTRpred will be visualized along the chromsome.

plotChromoMap(pred, genome.file, centromere_starts, ...)

Arguments

pred
LTRpred.tbl generated with LTRpred.

genome.file
a file path to the genome assembly file in fasta format for which chromosomes shall be visualized.

centromere_starts
a numeric vector containing the start coordinates of the centromers (for all chromosomes in genome.file).

References

https://cran.r-project.org/web/packages/chromoMap/vignettes/chromoMap.html

Examples

test_genome <- system.file("Hsapiens_ChrY.fa", package = "LTRpred")
test_pred <- LTRpred::read.ltrpred(system.file("Hsapiens_ChrY_LTRpred_DataSheet.tsv", package = "LTRpred"))
test_centromere_starts <- 55000
# generate visualization
LTRpred::plotChromoMap(test_pred, test_genome, test_centromere_starts)

#> ********************************** __ __ ************
#> ** __**|__ * __* __ * __ __ * __ *|  |  |* __ * __ **
#> **|__**|  |*|  *|__|*|  |  |*|__|*|  |  |*|_ |*|__|**
#> ***********************************************|   **
#> *************** by Lakshay Anand ********************
#> OUTPUT: 
#> Number of Chromosome sets: 1 
#> Number of Chromosomes in set  1 : 1 
#> Processing data.. 
#> Number of annotations in data set  1 : 21