Plot correlation between #cluster members and any other variable
Source: R/plotCluster.R

Visualize the correlation between the Cluster Number generated by LTRpred and any other variable and color each putative LTR transposon by it's e.g. superfamily name.

plotCluster(data, y = "CG_3ltr_rel", color.by = "fam",
  quality.filter = FALSE, sim = 70, n.orfs = 1, chop.names = TRUE,
  xlab = "# CLUSTER MEMBERS", ylab = "", main = "",
  legend.title = "Protein Domain")

Arguments

data
data.frame returned by LTRpred.

y
column name in data (lazy evaluation) that shall be correlated with cn_3ltr (copy number of predicted 3' LTR).

color.by
color points by a column name in data (lazy evaluation).

quality.filter
shall a quality filter be applied to reduce false positives?

sim
filter minimum sequence similarity between the LTRs.

n.orfs
filter for minimum number of orfs.

chop.names
shall superfamily names be chopped?

xlab
x-axis label.

ylab
y-axis label.

main
main text.

legend.title
legend text.