Plot the age distribution of predicted LTR transposons

This function visualizes the age distribution of predicted LTR transposons generated with LTRpred.

The age pf LTR transposons is defined by the sequence similarity between it's 3' and 5' LTR. Evolutionary young (recent) LTR transposons tend to have very similar LTRs (up to 100% sequence similarity), whereas evolutionary older LTR transposons tend to have less similar LTRs.

plotLTRSim(
  data,
  type = "hist",
  stack.fill = "protein_domain",
  similarity.bin = 2,
  min.sim = 70,
  quality.filter = TRUE,
  n.orfs = 1,
  xlab = "LTR % Similarity",
  ylab = "Frequency",
  main = "LTR Age Distribution",
  legend.title = "LTR Similarity"
)

Arguments

data
the data.frame generated by LTRpred.

type
type of histogram. Either normal histogram (type = "hist") or stacked histogram (type = "stack", see also stack.fill). If type = "stack" is specified then

stack.fill
a character string specifying the variable by which the bar plot shall be stacked.

similarity.bin
resolution of similarity binning. E.g. binning 98%-100% into 0.5% intervals would be similarity.bin = 0.5. Default is similarity.bin = 0.5.

min.sim
minimum similarity between LTRs that can shall be considered for visualization. All elements not fulfilling this similarity threshold are filtered out. Default is min.sim = 2.

quality.filter
shall false positives be filtered out as much as possible or not. See Description for details.

n.orfs
minimum number of ORFs detected in the putative LTR transposon.

xlab
x-axis label.

ylab
y-axis label.

main
main text.

legend.title
legend text.

Details

This way of visualizing the age distribution of LTR transposons allows users to examine the rate of recent transposition events in extant organisms.

LTR similarity values are binned in intervals (per default 2.5% intervals) which can be modified by the similarity.bin argument.

See also

Author

Hajk-Georg Drost

Examples

if (FALSE) {
# run LTRpred for A. thaliana
Ath.Pred <- LTRpred(genome.file = "TAIR10_chr_all.fas")
# visualize the age distribution of predicted  A. thaliana LTR transposons
PlotLTRAge(Ath.Pred)
}