Genome Size vs. Protein Space
Source: R/plotLTRSpace.R

Compute and visualize the correlation between genome size vs. Protein Space.

plotLTRSpace(genome.matrix, cor.method = "spearman", label.organism = TRUE,
  smooth.method = "glm", se = TRUE,
  xlab = "LTR transposon content in Mega [bp]",
  ylab = "Genome size in Mega [bp]", main = "", alpha = 0.3,
  arrow_lab = FALSE, cl.analysis = FALSE, cl.centers = NULL,
  cl.method = "euclidean", cl.nstart = 100, cl.iter.max = 10000,
  sim.matrix = NULL, text.size = 18, label.size = 3,
  check.overlap = TRUE)

Arguments

genome.matrix
Genome matrix retruned by LTRpred.meta.

cor.method
method to correlate genome size versus protein space.

label.organism
shall organism names be drawn in the figure?

smooth.method
method to add a smoothed conditional mean (see ggplot2::geom_smooth()). Users can choose from auto, glm, lm, or loess. If no smoothing function shall be applied, users can specify smooth.method = 'NULL'.

se
shall standard error of the smoothing model (grey area) be drawn or not.

xlab
x-axis label.

ylab
y-axis label.

main
main text.

alpha
transparency of confidence interval of the smoothing function: between \[ 0,1 \].

arrow_lab
whether or not data points shall be labelled with arrows (ggrepel).

cl.analysis
logical value indicating whether or not cluster analysis of the sim.matrix returned by LTRpred.meta shall be performed (see Details sections).

cl.centers
number of expected clusters or a set of initial (distinct) cluster centres.

cl.method
distance measure to perform cluster analysis. Options are "euclidean", "maximum", "manhattan", "canberra", "binary", "pearson" , "abspearson" , "abscorrelation", "correlation", "spearman" or "kendall".

cl.nstart
If cl.centers is a number, number of random sets that shall be chosen.

cl.iter.max
maximum number of iterations for cluster analysis.

sim.matrix
Similarity matrix retruned by LTRpred.meta.

text.size
size of the labels in ggplot2 notation.

label.size
size of the dot-labels.

check.overlap
shall overlaps of labels be avoided or not.

Details

In the first step, users should have generated LTR retrotransposon annotations for several species using the LTRpred.meta function. The output of LTRpred.meta is a folder storing the annotation files for all species of input species. In addition, LTRpred.meta generates two files: *_SimilarityMatrix and *_GenomeInfo .