Plot the median cluster copy number distribution of putative LTR transposons for several species

Visualize the correlation between LTR transposon age (measured in LTR similarity) and the median cluster copy number of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element for each species stored in the input data.

plot_copynumber_species(data, element.type = "full_retrotransposon",
  plot.type = "boxplot", similarity.bin = 2, min.sim = 70,
  quality.filter = FALSE, n.orfs = 0, include.unique = FALSE,
  xlab = "LTR % Similarity", ylab = "Median Copy Number", main = "",
  legend.title = "Similarity between LTRs", arrow_lab = FALSE,
  y.ticks = 10)

Arguments

data
the data.frame generated by LTRpred.

element.type
which width shall be visualized. Full retrotransposon width or only LTR element cluster copy number: choose element.type = "full_retrotransposon" or element.type = "ltr_element". In case element.type = "ltr_element" is chosen then the width of the left LTR element is used to visualize width vs. seq. similarity between the LTRs.

plot.type
type of visualization: either plot.type = "boxplot" or plot.type = "violin".

similarity.bin
resolution of similarity binning. E.g. binning 98%-100% into 0.5% intervals would be similarity.bin = 0.5.

min.sim
minimum similarity between LTRs that can shall be considered for visualization.

quality.filter
shall false positives be filtered out as much as possible or not. See Description for details.

n.orfs
minimum number of ORFs detected in the putative LTR transposon.

include.unique
shall unique elements be visualized together with the clusters?

xlab
label of x-axis.

ylab
label of y-axis.

main
title label.

legend.title
title of legend.

arrow_lab
whether or not data points shall be labelled with arrows (ggrepel).

y.ticks
number of ticks at the y-axis. Default is y.ticks = 10.

Details

This function visualizes the correlation between LTR transposon age (measured in LTR similarity) and the cluster copy number of the corresponding LTR retrotransposon (the complete sequence) or the width of the left LTR element. Using this visualization approach, different classes of LTR retrotransposons can be detected due to their width and age correlation.

Examples

# NOT RUN {
# }