Plot positions of predicted retrotransposons along chromosomes

The positionas of LTR retrotransposons predicted with LTRpred will be visualized along the chromsome.

plot_element_distr_along_chromosome(pred, genome.file, centromere_starts, ...)

Arguments

pred
LTRpred.tbl generated with LTRpred.

genome.file
a file path to the genome assembly file in fasta format for which chromosomes shall be visualized.

centromere_starts
a numeric vector containing the start coordinates of the centromers (for all chromosomes in genome.file).

...
additional arguments that shall be passed to the visualization function autoplot.

Examples

if (FALSE) {
test_genome <- system.file("Hsapiens_ChrY.fa", package = "LTRpred")
test_pred <- LTRpred::read.ltrpred(
system.file("Hsapiens_ChrY_LTRpred_DataSheet.tsv",
             package = "LTRpred"))
test_centromere_starts <- 55000
# generate visualization
LTRpred::plot_element_distr_along_chromosome(test_pred, test_genome, test_centromere_starts)
}