Read output of ORFpred
Source: R/read.orfs.R

This function reads the output of the ORFpred function and stores the sequence id and number of predicted ORFs in a data.frame object.

read.orfs(input.file)

Arguments

input.file
fasta file generated by ORFpred.

Value

A data.frame object storing the seq.id, orfs (number of predicted ORFs), start, and end of the predicted LTRs.

Details

The file generated by ORFpred is parsed by this function and returned as data.frame object.

See also

Examples

# read an example prediction file generated by PredictORFs()
ORFPred <- read.orfs(system.file("nt.fa",package = "LTRpred"))

head(ORFPred)

#> # A tibble: 6 x 2
#>   seq.id                                  orfs
#>   <chr>                                  <dbl>
#> 1 1_CHROMOSOME_dumped__10696719_10701589    2.
#> 2 1_CHROMOSOME_dumped__11148084_11152830    1.
#> 3 1_CHROMOSOME_dumped__11284903_11295619    2.
#> 4 1_CHROMOSOME_dumped__11339006_11344662    1.
#> 5 1_CHROMOSOME_dumped__11346530_11355061    2.
#> 6 1_CHROMOSOME_dumped__11697950_11705480    1.