Import the output of LTRharvest or LTRdigest

This function imports the output files generated by LTRharvest or LTRdigest.

read.prediction(
  gff.file = NULL,
  tabout.file,
  program = "LTRdigest",
  ltr.fasta = NULL,
  inner.seq.fasta = NULL,
  data = NULL,
  similarity.bin = 2,
  min.sim = NULL
)

Arguments

gff.file
path to GFF file generated by either LTRharvest, LTRdigest, or LTRpred.

tabout.file
path to *_tabout.csv file generated by either LTRdigest or or LTRpred.

program
prediction program used to generate the corresponding GFF file. Either program = "LTRharvest", program = "LTRdigest", or program = "LTRpred".

ltr.fasta
fasta file generated by either LTRdigest or or LTRpred storing the sequences of the predicted LTR retrotransposons (e.g. *-ltrdigest_complete.fas).

inner.seq.fasta
path to the *_BetweenLTRSeqs.fsa fasta file returned by LTRharvest in caste program = "LTRharvest" is specified.

data
the data.frame returned by either LTRharvest, LTRdigest, or LTRpred that can be used to modify for example similarity.bin and min.sim within this table.

similarity.bin
resolution of similarity binning. E.g. binning 98%-100% into 0.5% intervals would be similarity.bin = 0.5.

min.sim
minimum similarity between LTRs that can shall be considered for visualization.

Author

Hajk-Georg Drost

Examples

if (FALSE) {
gff.file <- system.file("TAIR10_chr_all_LTRdigestPrediction.gff", package = "LTRpred")
tabout.file <- system.file("TAIR10_chr_all-ltrdigest_tabout.csv" ,package = "LTRpred")
LTRfile <- read.prediction(gff.file,tabout.file, program = "LTRdigest")

# look at results
str(LTRfile$ltr.retrotransposon)

}