Clean the initial Repbase database for BLAST
Source: R/repbase.clean.R

Clean the headers of the Repbase fasta files so that headers can be used to create a blast-able database.

repbase.clean(repbase.file, output.file)

Arguments

repbase.file
fasta file storing the corresponding Repbase annotation, e.g. athrep.ref.

output.file
name/path of the cleaned Repbase annotation file.

Details

The Repbase database can be downloaded after registration at http://www.girinst.org/repbase/. The corresponding files as they are however, cannot be converted into a blast-able database. Hence, a pre-filtering step is neccessary to be able to use this database with the e.g. repbase.query function.

References

http://www.girinst.org/repbase/

Examples

# NOT RUN {
# PreProcess Repbase: A thaliana
# and save the output into the file "Athaliana_repbase.ref"
repbase.clean(repbase.file = "athrep.ref",
              output.file  = "Athaliana_repbase.ref")

# }